Use of a NoopMutex
